+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 ## 0.1.0
 
-* added contract abis and generated web3 contracts
-*
+* contract abis and generated web3 contracts
+* publish ballot spec